REFACTOR: move SAT variables used for dpll to dpll, SAT variables expected to be used in other places to utils
