--- satsolver/dpll/pure_literal_elimination.py
@@ -1,5 +1,0 @@
-def find_pure_literal():
-  pass
-
-def pure_literal_elimination():
-  pass
